+++ internal/zero/controller/usagereporter/usagereporter_test.go
@@ -1,4 +1,4 @@
-package controller
+package usagereporter
 
 import (
 	"context"
